+++ src/main/java/com/assesment/inventory_management/service/ProductService.java
@@ -13,6 +13,7 @@ import com.assesment.inventory_management.repository.ProductRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -48,7 +49,7 @@ public class ProductService {
         return convertToResponseDTO(product);
     }
 
-    public List<ProductResponse> getAllProducts(Long categoryId, Integer minPrice, Integer maxPrice) {
+    public List<ProductResponse> getAllProducts(Long categoryId, BigDecimal minPrice, BigDecimal maxPrice) {
         List<Product> products;
 
         if (categoryId != null && minPrice != null && maxPrice != null) {
